FAT: Implement trunc_file. Move read_sync and write_sync to a different file.
